Cleanup cleanup installation-callback
