+++ apps/app/src/app/api/installation-callback/schema.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const schema = z.object({
+  code: z.string(),
+  state: z.string().nullish(),
+  installationId: z.coerce.number().nullish(),
+  setupAction: z.enum(["install", "request"]),
+});
